Accompany list of commands by default key bindings
While changing a couple of key bindings, I had to switch a lot between
the "Default key bindings" section and the list of commands, to check
which keys were already taken.

In my opinion, it would be easier for the user if the list of commands
were accompanied by the default key bindings, so they can just be copied
and pasted into the configuration file.